remove helper functions to separate python file utilities.py
